+++ zebra-rpc/src/methods/types.rs
@@ -1,6 +1,5 @@
 //! Types used in RPC methods.
 
-pub mod default_roots;
 pub mod get_block_template;
 pub mod get_blockchain_info;
 pub mod get_mempool_info;